+++ contracts/TokenTRC20.sol
@@ -12,8 +12,8 @@ contract TokenTRC20 {
 
     uint256 public totalSupply;
 
-    mapping (address => uint256) public balanceOf;
-    mapping (address => mapping (address => uint256)) public allowance;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
     // This generates a public event on the blockchain that will notify clients
     event Transfer(address indexed from, address indexed to, uint256 value);
@@ -21,7 +21,11 @@ contract TokenTRC20 {
     // This notifies clients about the amount burnt
     event Burn(address indexed from, uint256 value);
 
-    event Approval(address indexed _owner, address indexed _spender, uint _value);
+    event Approval(
+        address indexed _owner,
+        address indexed _spender,
+        uint256 _value
+    );
 
     /**
      * Constructor function
@@ -33,23 +37,32 @@ contract TokenTRC20 {
         string memory tokenName,
         string memory tokenSymbol
     ) public {
-        totalSupply = initialSupply * 10 ** uint256(decimals);  // Update total supply with the decimal amount
-        balanceOf[msg.sender] = totalSupply;                // Give the creator all initial tokens
-        name = tokenName;                                   // Set the name for display purposes
-        symbol = tokenSymbol;                               // Set the symbol for display purposes
+        totalSupply = initialSupply * 10**uint256(decimals); // Update total supply with the decimal amount
+        balanceOf[msg.sender] = totalSupply; // Give the creator all initial tokens
+        name = tokenName; // Set the name for display purposes
+        symbol = tokenSymbol; // Set the symbol for display purposes
     }
 
     /**
      * Internal transfer, only can be called by this contract
      */
-    function _transfer(address _from, address _to, uint _value) internal {
-
+    function _transfer(
+        address _from,
+        address _to,
+        uint256 _value
+    ) internal {
         // Check if the sender has enough
-        require(balanceOf[_from] >= _value, "TokenTRC20: the balance must be greater than value");
+        require(
+            balanceOf[_from] >= _value,
+            "TokenTRC20: the balance must be greater than value"
+        );
         // Check for overflows
-        require(balanceOf[_to].add(_value) >= balanceOf[_to], "TokenTRC20: the balance must be greater than value");
+        require(
+            balanceOf[_to].add(_value) >= balanceOf[_to],
+            "TokenTRC20: the balance must be greater than value"
+        );
         // Save this for an assertion in the future
-        uint previousBalances = balanceOf[_from].add(balanceOf[_to]);
+        uint256 previousBalances = balanceOf[_from].add(balanceOf[_to]);
         // Subtract from the sender
         balanceOf[_from] = balanceOf[_from].sub(_value);
         // Add the same to the recipient
@@ -67,7 +80,10 @@ contract TokenTRC20 {
      * @param _to The address of the recipient
      * @param _value the amount to send
      */
-    function transfer(address _to, uint256 _value) public returns (bool success) {
+    function transfer(address _to, uint256 _value)
+        public
+        returns (bool success)
+    {
         _transfer(msg.sender, _to, _value);
         return true;
     }
@@ -81,8 +97,15 @@ contract TokenTRC20 {
      * @param _to The address of the recipient
      * @param _value the amount to send
      */
-    function transferFrom(address _from, address _to, uint256 _value) public returns (bool success) {
-        require(_value <= allowance[_from][msg.sender], "TokenTRC20: value cannot be greater than the allowed value");
+    function transferFrom(
+        address _from,
+        address _to,
+        uint256 _value
+    ) public returns (bool success) {
+        require(
+            _value <= allowance[_from][msg.sender],
+            "TokenTRC20: value cannot be greater than the allowed value"
+        );
         allowance[_from][msg.sender] = allowance[_from][msg.sender].sub(_value);
         _transfer(_from, _to, _value);
         return true;
@@ -96,14 +119,14 @@ contract TokenTRC20 {
      * @param _spender The address authorized to spend
      * @param _value the max amount they can spend
      */
-    function approve(address _spender, uint256 _value) public
-    returns (bool success) {
+    function approve(address _spender, uint256 _value)
+        public
+        returns (bool success)
+    {
         allowance[msg.sender][_spender] = _value;
         emit Approval(msg.sender, _spender, _value);
         return true;
     }
-
-
 
     /**
      * Destroy tokens
@@ -113,7 +136,10 @@ contract TokenTRC20 {
      * @param _value the amount of money to burn
      */
     function burn(uint256 _value) public returns (bool success) {
-        require(balanceOf[msg.sender] >= _value, "TokenTRC20: the balance must be greater than value");
+        require(
+            balanceOf[msg.sender] >= _value,
+            "TokenTRC20: the balance must be greater than value"
+        );
         balanceOf[msg.sender] = balanceOf[msg.sender].sub(_value);
         totalSupply = totalSupply.sub(_value);
         emit Burn(msg.sender, _value);
@@ -128,8 +154,14 @@ contract TokenTRC20 {
      * @param _from the address of the sender
      * @param _value the amount of money to burn
      */
-    function burnFrom(address _from, uint256 _value) public returns (bool success) {
-        require(balanceOf[_from] >= _value, "TokenTRC20: the balance must be greater than value");
+    function burnFrom(address _from, uint256 _value)
+        public
+        returns (bool success)
+    {
+        require(
+            balanceOf[_from] >= _value,
+            "TokenTRC20: the balance must be greater than value"
+        );
         require(_value <= allowance[_from][msg.sender]);
         balanceOf[_from] = balanceOf[_from].sub(_value);
         allowance[_from][msg.sender] = allowance[_from][msg.sender].sub(_value);
